pyproject.toml
[project]
name = "mcg"
version = "0.1.0"
description = "Generalizable Black-box Adversarial Attack with Meta Learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "advertorch @ git+https://github.com/cylix-reproduction/advertorch",
    "argparse>=1.4.0",
    "astunparse>=1.6.3",
    "cma>=4.3.0",
    "dacite>=1.8.1",
    "h5py>=3.14.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "punq>=0.7.0",
    "torch-dct>=0.1.6",
    "torchattacks>=3.5.1",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
cpu = ["torch>=2.5.0", "torchvision>=0.20.0"]
cu118 = ["torch>=2.5.0", "torchvision>=0.20.0"]
cu126 = ["torch>=2.5.0", "torchvision>=0.20.0"]
cu128 = ["torch>=2.5.0", "torchvision>=0.20.0"]

# PyTorch indexes.
#
# The following indexes are used to install a specific version of PyTorch. They're only available when `uv` is used.

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu118" },
    ],
    [
        { extra = "cpu" },
        { extra = "cu126" },
    ],
    [
        { extra = "cpu" },
        { extra = "cu128" },
    ],
    [
        { extra = "cu118" },
        { extra = "cu126" },
    ],
    [
        { extra = "cu118" },
        { extra = "cu128" },
    ],
    [
        { extra = "cu126" },
        { extra = "cu128" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# Disable Pyright (BasedPyright) since this project is forked.
[tool.basedpyright]
typeCheckingMode = "off"

[dependency-groups]
dev = [
    "tqdm>=4.67.1",
]
